Sunglasses_Prediction/requirements.txt
flask==2.3.2
flask-pymongo==2.3.0
scikit-learn==1.3.0
pillow==10.0.0   
requests==2.29.0 